Improve error handling of process_ensembles.py
Any error from chs_create_initial_masters.py is now logged, previously
it was dropped (so you just got a message that an error had occurred,
not the text of the error).

Cleaned up the processing code so that the mapping from ensemble to
stack is now a required file rather than having a fall-back version
from testing hard-coded in chs_create_initial_masters.py.