JS9 displays a particular component, not the stack
This moves back to the original scheme, whereby JS9 displays data
for a particular component, rather than stack. This means that the
band can be correct for that component. The band is now included,
as well as the component number, in the label for the JS9 display
area.

At present the choice of band is the *only* difference in the JS9
display, as there is no attempt to separate out the component hull
from any other hulls in that stack.