Track more information for stack-level hulls
Include information on likelihood, whether to include in the centroid
calculation, whether the stack went to SVD QA.

Change the master-id numbering scheme to hopefully be more robust. It
is not clear it really makes a realistic difference, but it does change
the numbering for some ensembles. This means that user-analysis
prior to this change is not guaranteed to match by master id.