Avoid warning when re-loading JS9
The code has an implicit assumption that we only have a single
JS9 image for a stack for the masterhull page. This is actually
not true:

  a) can load multiple copies simultaneously
  b) can exit a JS9 and then re-load

We want to allow case b, and case a could be useful. Simple
testing suggests that it is okay, but not 100% clear.

This commit just removes a warning that would be displayed when
the assumed invariant was broken.