Basic working version of chs_complete_ensemble.py
This only supports "simple" ensembles - those where the proposed hulls
are accepted or deleted. No manual adjustment (either of the hull or
the components that form each master) are allowed.

Much of the logic in chs_complete_ensemble.py needs to be moved to
chs_utils and merged with chs_finalize_ensemble.