Fix QA UI when an ensemble had a deleted master hull
The fix was to explicitly convert a dictionary to JSON syntax, via
tojson, since the -1 keyword needs to be protected and it wasn't being
done so with the default conversion.

In tracking this down I have moved some of the "larger" blobs of JSON
to there own API. It could be added to the general "master-hull"
status page, since this would have saved a lot of repeated work, but
for now have separate endpoints and requests.